Replace owl:sameAs with schema:sameAs in most places
LicenseDB used to claim that

<https://licensedb.org/id/BSD-3-Clause> owl:sameAs <http://opensource.org/licenses/BSD-3-Clause> .

This was incorrect:

<https://licensedb.org/id/BSD-3-Clause> is a free software license
<https://licensedb.org/id/BSD-3-Clause.html> is a web page about a free software license
<http://opensource.org/licenses/BSD-3-Clause> is a web page about a free software license

Obviously "something" and "a web page about something" are not the same thing [1].  For
these uses owl:sameAs has been replaced with schema:sameAs, which states that its object is
the "URL of a reference Web page that unambiguously indicates the item's identity. E.g. the
URL of the item's Wikipedia page, Freebase page, or official website."

This change has not been made for the creative commons licenses, which use the URL of the
web page about a creative commons license as an identifier for the license itself in their
own published RDF data.

[1] See also: https://en.wikipedia.org/wiki/HTTPRange-14