Add the split out fedora.json license as individual YAML files
This is the result of running json2yaml.py on fedora.json.  Licenses
that carry a Fedora abbreviation are written to ./fedora_abbrev with
the abbreviation as the filename plus a '.yml' extension.

Licenses with only an SPDX abbreviation and no Fedora abbreviation go
to ./spdx_abbrev and use the SPDX name for the file plus a '.yml'
extension.

All other licenses go to ./other and the filename is the main JSON
file key for the license with '.yml' on the end.

Any filename that ends up with a slash in it has the slash changed to
an underscore.  So far that has hit these licenses from fedora.json:

    C/Migemo License
    Microsoft's Shared Source CLI/C#/Jscript License

Those are in the other/ subdirectory.