Cleaning up the license data files
Break out the JSON file in to one file per license and then group
those in to the data/ tree.  The subdirectories in data/ are by
organization, company, or project.  The purpose is to make it easier
to maintain.  The scripts we will add later will traverse this tree to
collect all entries to generate the output files, so the structure of
data/ is freeform for directories.  License files must be *.yaml files
and carry the contents of the top level TEMPLATE.yaml file.

Other changes:

* The id entries have been removed.

* The license_text field has been removed.

* If a field was empty in the JSON file, it does not exist in the YAML
  file.

* All single quotes have been removed.

There is more cleanup and auditing work to do, but the main structure
is here now.

TEMPLATE.yaml explains the valid fields in a license file.

Signed-off-by: David Cantrell <dcantrell@redhat.com>