curation-generator: Add version ranges for the aspnetcore curations
The repository was reorganized several times, so the existing curations
without versions did only work for some versions of the packages. To fix
this, check the paths for all projects for all release tags and generate
curations with version ranges. Leave the last range open ended so that
the curation works for future releases, until the repository is
reorganized again.

The curations will be updated in a separate commit for a nicer diff.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>