Rename tsv2toml.py to tsv2toml-good.py; add tsv2toml-bad.py
The source spreadsheets used to get the data to import have a good
license list and a bad license list, but the formats of the
spreadsheets are different.  So now there are two import scripts.  One
for the good license spreadsheet and one for the bad.  Update the
README with my own notes gathering the latest data.