Add tools/validate-spec.py script and update the Makefile
You can now type "make validate" and it will run toml-validator to
ensure all the TOML files use valid syntax.  Then it will run
tools/validate-spec.py which is our rules as defined in TEMPLATE.toml.
Anything not in compliance is an error.  The idea is to use this
script to validate pull requests.  Exit code of 0 means everything is
good.

NOTE: Any changes to TEMPLATE.toml requires an update to
validate-spec.py.  The script does not read TEMPLATE.toml to get the
rules to enforce.  TEMPLATE.toml is meant for human reading.

Signed-off-by: David Cantrell <dcantrell@redhat.com>