Add the presumed 2014 version of the Microsoft .NET Library EULA as not-allowed
This license was identified by ScanCode Toolkit in a Red Hat review of a Red Hat product
package of dotnet in 2019. I believe the package maintainer of the Red Hat product version
was also the maintainer of corresponding upstream Fedora package(s). I cannot practically
verify whether the ScanCode reference text is an accurate match to whatever was in the
package that ScanCode flagged as a match. The URL that I believe ScanCode recorded for this
license has a somewhat different version today
(http://www.microsoft.com/web/webpi/eula/net_library_eula_enu.htm). Based on some cursory
Wayback Machine research I am guessing that the version here was in place at least as far back
as 2014. Since this license might have been present in a Fedora package in the recent past and
it (or a similar license) foreseeably might be identified in a Fedora package of dotnet in the
future, it seems justifiable to add this to the Fedora License Data repository.

To be clear, this license was presumably included in error by the upstream dotnet maintainers.

Signed-off-by: Richard Fontana <rfontana@redhat.com>