evaluator.rules: Ignore "stand-alone" exceptions
If the license ID (not the exception ID) of an
`SpdxSingleLicenseExpression` refers to an exception, ignore it in order
to silence the corresponding UNHANDLED_LICENSE violation.

Such "stand-alone" exceptions are flagged in case the license scanner
was unable to associate that exception with the corresponding license.
In that case the scanner reports the license ID and exception ID
separately. As exceptions only weaken the obligations ignoring the
exceptions does not make the rules less strict and thus should be fine.

It removes distraction from the report, as errors for unassociated
exceptions are roughly speaking flagged twice: once for the license and
once for the exception.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>