evaluator.rules: Add handling for 'generic' licenses
The generic license category contains license IDs which do not represent
an exact license, but triggers which only indicate that the tooling has
found something which may be a license.

Generic licenses do require investigation, as (if the finding is valid)
a dedicated license ID needs to be created. As some of these triggers
match quite broadly huge amount of manual effort can be created.
Requiring the investigation of all generic triggers in fact only leads
to moving the focus from working on the most important things towards
working mostly on irrelevant details with the result of not getting
anything done anymore.

So, introduce a hard-coded list of `ignoredLicenses` containing license
IDs which should not be flagged, to avoid that problem.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>