fix(tool/curations): Drop an obsolete category override entry
The issue has been fixed upstream and the update has been published,
see https://scancode-licensedb.aboutcode.org/ms-cla.json.

Keep the mechanism to make it easy to override categories in the
future. Also there may arise need to make the override categories
configurable in the future which means there is likely need for this
mechanism.

Signed-off-by: Frank Viernau <frank_viernau@epam.com>