Add public domain notices from the tzdata data files
From here:  https://github.com/eggert/tz

Capture the public domain notices that appear in the data files.  The
source code in the project is also public domain but the comments are
in a different format.  I did not capture that here because Fedora
does not ship the programs this code generates.

Signed-off-by: David Cantrell <dcantrell@redhat.com>