build: Add a minimum Gradle framework for editing ORT scripts
In order to get auto-completion for custom ORT scripts working, the
scripting definitions need to be known. To make them known, a Gradle
project which pulls in the respective dependencies is required. Add the
required Gradle files so that opening a Gradle build file in IntelliJ or
opening the folder in Fleet [1] will load the scripting definitions, and
subsequently opening *.rules.kts, *.notifications.kts, or
*.how-to-fix-text-provider.kts files will make auto-completion work.

[1]: https://www.jetbrains.com/fleet/

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>